Camera implementation

+++ src/game/Vector2D.cpp
@@ -4,7 +4,7 @@ Vector2D::Vector2D() : x(0.0f), y(0.0f)
 {
 }
 
-Vector2D::Vector2D(float x, float y) : x(x), y(y)
+Vector2D::Vector2D(int x, int y) : x(x), y(y)
 {
 }
